backend/services/requirements.txt
aiobotocore==2.15.2
aiosqlite==0.20.0
asyncpg==0.30.0
curlify2==2.0.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
httptools==0.6.4
httpx==0.27.2
ipykernel==6.29.5
Jinja2==3.1.4
multipledispatch==1.0.0
pipdeptree==2.24.0
psycopg2==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
SQLAlchemy==2.0.36
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
